Made Alot Of Changes : Leaderboard UI, Ofline Earnings, Powers, And So Much More...

+++ FunnyMoneyClicker/Assets/Scripts/ClickerManager.cs
@@ -38,7 +38,7 @@ public class ClickerManager : MonoBehaviour
 
         displayedMoney += (SaveDataController.currentData.moneyCount - displayedMoney) * Time.deltaTime * 4.0;
 
-        if (Mathf.Abs((float)(SaveDataController.currentData.moneyCount - displayedMoney)) < 0.01f)
+        if (Mathf.Abs((int)(SaveDataController.currentData.moneyCount - displayedMoney)) < 0.01f)
         {
             displayedMoney = SaveDataController.currentData.moneyCount;
         }
